requirements.txt
numpy==1.14.0
pandas==0.22.0
six>=1.11<1.12
lazy-object-proxy>=1.3<1.4
scikit-learn>=0.19<0.20
scipy>=1.0<1.1
docopt>=0.6<0.7
colorama>=0.3<0.4
tables>=3.4<3.5
umap-learn>=0.2<0.3
ggplot>=0.11<0.12
